Coordinate system changed.
Now the projection plane for the two emisphere of the LL images
is vertical with its normal direction pointing forward.
The coordinate system has changed so that x-axis points to the
right, the y-axes points down and the z-axis points forward. This
may be convenient to compare orientation and translation during
SFM.

The IntrinsicMatrix has been corrected by introducing u0 and v0,
the coordinate of the principal point.

The epipolar constraint error is little but there is no meaningful
output yet.