LL points are projected again to make SFM works.
Estimating essential matrix works both with normal and LL images.
In order to make it work with omnidirectional cameras, I project
the image points of the unitary sphere on a plane with a
perspective projection. The focal length of this new projection is
1 and the center of projection must be the same point of the
original LL image.

This fix allows me to compute the correct essential matrix but the
FOV of the input image is reduced drastically (this script is
equivalent to extract a perspective projection out of a panoramic
image).

Now I am going to move on and see if the essential matrix
estimation works with this projection approach when the center of
projection is different from the point used by panoramic rendering.