SFM seems to work. Not completely tested yet.
A step forward was made when I changed the image inputs. Some
points whose projections were almost centered in the perspective
projection output were triangulated wrong. I think this is due to
a poor choiche of the distance between the two consecutive images
and the long distance from the camera of the tracked points.
If the cameras are brought closer to a foreground object, the
reprojection results seems better. I still have to do proper tests
with an ad-hoc synthetic model (possibly coupled with ground truth
data).

I made some fixes in the sfm script when using the triangulate
function to triangulate points in a pair of views. I am not going
to keep this function though because triangulateFromMultiViews can
deal with different numbers of points tracked in each view (which
is the case when dealing with more than two views).

The function to compute the relative scale return the median value
for the scale computed among the points whose reprojectionError is
below a given threshold.