last improvements before dropping windowedTracking
There is no benefit in using windowed trackging as it is now.
Since the tracked point has to be visible in consecutive views and
it is added to a connections only if it is visible in every view of
the window.

It may be useful to select only the longer tracks, but then it may
be implemented in a simpler way (maybe just by using findTracks and
then choose the longer tracks only). It may be useful if we want to
track points in consecutive views but, again, findTracks, toghether
with some clever cheks can bring us the same benefits with less
complexity.