New E estimating helper function w/out projection
The new function uses the point conversion obtained with
createPointsConversionTable which in turns does not use key points
projection (it divides all points with their z-component).

The function has some parameters that can be used to customize its
behavior (like when to discard key points that reside on the back).