New tracks finding function
This new function creates tracks based on the viewSet's matches,
then it tries to track each points in every window's view. If it
succeds, the track is checked for matching with initial every view
where the corresponding point appears (not implemented yet) and if
this last check is successfull, it adds the track to the viewSet in
form of a set of connections between the window's views.