Dynamic E estimation introduced
helperEstimateRelativePose improved with the so called Dynamic
estimation technique: if the number of inliers after the E estimation is
too low, E is computed again up to a maximum number of trials.

E evaluation script improved: some unused variables have been removed
and the inliers and matches numbers for each estimation are now saved in
the output file.