Dynamic E estimation improved.
We performed some test to decide what is the best metric to evaluate a
pose estimation and we opted for the
pointsForEEstimationCounter/pointsForPoseEstimationCounter ratio.
In order to decide whether the estimated pose is good enough or not, the
previous ratio has to be less than maxInliersRatio.

The metric has been evaluated based on the results of test4