viewWindow improved for frontal points only
The computeTracksAndCreateConnections has been improved in order to take
into account the indexPairs created by matchFeatures. These pairs are
edited by the helperEstimateRelativePose function which removes backward
points.
The test script for viewWindow works fine.

An IMPORTANT fix has been added to createPointsConversion. It should
create a correct list of frontal points now. This fix has to be tested
further in order to decide if it is really correct.