Back points fixed, new dynamic eval. Tested
The problem with persisting back points is now solved.
The fix has been tested and the results evaluated and compared with
repetions without back points.

The value validPtsFraction become meaningful after this fix and a new
dynamic evaluation metric has been implemented (which considers both the
total/frontal matches ratio and validPtsFraction, the ratio of inliers
points that project in front of both cameras)