E estimation function improved for further tests
The helperComputeRelativePose function has been improved in order to
allow for automatic test and the loop-ed version has to be preferred.

Added some utility function to plot bar plot with confidence intervals
and to compute confidence intervals.

computeRelativeScaleFromGroundTruth improved to take into account the
fact that the estimated location vector norm is not always 1.

ComputePoseError has a new input argument to ask this function either to
compute relative errors (available for location error only) or not.

A comment added to track computation testing script.