Error fixing with CI, improved E est. and plotting
The CI computing function had an error in the formula. Plus it has been
improved to work with a single repetition (it gives a CI equal to 0).

helperEstimateRelativePose is more robust now because it always compute
a pose estimation. It did not do so when the number of inliers was equal
to 0 and the iteration limit was reached.

A default argument in plotting function has been added. The series' name
is no more mandatory.