New seq filter implemented. Not tested yet
The new filter has one mode only. It computes the distances between
matching features in two images, takes a specific quantile and computes
the median. If it is above a threshold, the frame is used for pose
estimation, otherwise it is discarded.

The main function now converts to gray scale only the images that it is
going to use (not every image in the data folder).