Important fix to bdlAdj fixed views
There was an error in the creation of the fixed view list. This is one
of the parameters of the bundle adjustment routine. The problem caused
incorrect results with some experiments. The effect was that just the
first view of the window remained fixed instead of both the first and
second view in window.